+++ src/main/java/pl/pop/interview/master/account/AccountRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Account findByEmail(String email);
+public interface AccountRepository extends JpaRepository<Account, String> {
 }
